add script to generate config reference tables
For the Jenkins package, we maintain a configuration reference with our
documentation. Up to this point, we've generated Markdown tables with
configuration properties by hand, but the package is at the point where
it's no longer feasible for someone to do this by hand.

This commit introduces a script that walks the Universe package
repository and builds a single Markdown file with a configuration
reference. A user can run this script, get a single Markdown file in
their current working directory, and copy/paste the resulting table into
their documentation tool of choice.