add dummy environment variable
There's no way to know which options a user has selected, but each of
them has a trailing comma in marathon.json.mustache (since they're all
part of the larger `env` JSON object. So here we add a dummy environment
variable that should be the very last variable in the object, and it
*does not* have a trailing comma. Since trailing commas aren't permitted
in JSON, this avoids any syntax errors that might arise.