Adding a parameters option
Adding a missing section that gets rendered empty if the end user has not supplied any additional parameters.

This will allow customers/users to inject values as needed.

Removing necessity for "{{last}}" and a lame comma

Taking feedback, I agree, it was kind of junk to have a {{last}}.

This new way, there's always an identifier passed through and available for inspection to see that this container was fired up by marathon.

This identifies the source as well as the actual package it was launched from.

Adding config.json definition

This should now match up with the marathon-lb moustache file.

Correcting parameters

Now properly identifying the parameters as LABELS.

Changing the parameter name

As pointed out, these parameters are specific to the marathon-lb mesos module.

Corrected.

Correcting whitespace

My new stanza wasn't aligned.  Fixing.

Validated extra comma was a problem

Parsing and compacting the mustache template proved this additional comma was a problem.