enhances validations
With this commit we will be able to run a certain set of validation
only against the most recent revision of a package. This is helpful to
not dig in historic, non-changeable files for (now) unsupported features.

Two checks are provided with this feature:
Check for non-https links and
validate that assets hosted on downloads.mesosphere.com are in a specific sub-directory.
The later only applies to packages hosted by mesosphere and we should use these WARNINGS
to lean up our distribution.
Suggestion: Go trough package by package and fix what is needed in a new revision each.