Updates for Confluent Platform packages to latest release (3.2.2) (#1264)
* Updating packages for Confleunt 3.2.2

* Aligned Kafka health check default timings between config.json and *.mustache.

More specific tool-tip around the broker memory and disk settings
(based on feedback from customers that current settings are confusing).

* Properly constrain memory allocation for docker image.
For now, we don't allow the user to re-configure the JVM Heap allocation
for the Control Center service (which defaults to 3 GB).  So the
minimum memory for the service should be 4 GB to allow room for
the RocksDB instance.