Introduce the new dcos-core-cli package
It contains the core subcommands (marathon, job, service, etc.).

It is part of our CLI refactoring which removes the need for user to
switch to the correct CLI each time they attach to a different cluster.

The core commands are now tied to a DC/OS cluster, this is in line with
the dcos-enterprise-cli package.

https://jira.mesosphere.com/browse/DCOS_OSS-3731